In this iteration we made sure that trees that don't get submitted to the cloud don't get erased from local storage so that they can be tried again later.
In order to simulate being offline, use the Network tab in the Inspect tools. Under "No throttling there is an option to go "offline".
We found out that an API call that fails does not trigger the catch part of the try/catch because it comes back as a success but with a null response.

Signed-off-by: Daniel Mermelstein <danmermel@gmail.com>